requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9

redis==5.0.3

numpy==1.26.4
faiss-cpu==1.7.4
rank-bm25==0.2.2

sentence-transformers==2.6.1
transformers==4.39.3
accelerate==0.29.3

torch==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

groq==0.4.2

langchain==0.1.14
langchain-core==0.1.42
langchain-community==0.0.30
langchain-text-splitters==0.0.2

httpx==0.25.2

pypdf==4.2.0
pandas==2.2.2
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0

scikit-learn==1.4.2
scipy==1.12.0